requirements.txt
# Core
python-dotenv==1.0.1

# LangChain
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.16
langchain-ollama==0.3.2

# Embeddings
sentence-transformers==4.1.0

# Vector Database
chromadb==0.4.24

# PDF
pypdf==5.4.0

# Backend
fastapi==0.115.12
uvicorn==0.34.2

# Frontend
streamlit==1.45.1

# Protobuf (chromadb 0.4.24 compatibility)
protobuf==3.20.3